Remove old and update new check
