+++ public/components/health-check/services/check-kibana-settings.ts
@@ -1,19 +1,7 @@
-/*
- * Wazuh app - Check Kibana settings Service
- *
- * Copyright (C) 2015-2021 Wazuh, Inc.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * Find more information about this on the LICENSE file.
- *
- */
-
 import { AxiosResponse } from 'axios';
 import { GenericRequest } from '../../../react-services';
+import { WAZUH_METAFIELDS_SETTINGS } from '../../../../common/constants'
+import _ from 'lodash';
 
 type userValue<T> = { userValue: T };
 
@@ -25,11 +13,11 @@ type kbnSettings = {
 type responseKbnSettings = { settings: kbnSettings };
 
 export function checkKibanaSettings(removeMetaFields: boolean) {
-  removeMetaFields &&
-    getKibanaSettings()
-      .then(checkMetafieldSetting)
-      .then(updateMetaFieldsSetting)
-      .catch((error) => error !== 'Unable to update config' && console.log(error));
+  removeMetaFields && 
+  getKibanaSettings()
+  .then(checkMetafieldSetting)
+  .then(updateMetaFieldsSetting)
+  .catch((error) => error !== 'Unable to update config' && console.log(error));
 }
 
 async function getKibanaSettings(): Promise<responseKbnSettings> {
@@ -39,12 +27,13 @@ async function getKibanaSettings(): Promise<responseKbnSettings> {
 
 async function checkMetafieldSetting({ settings }: responseKbnSettings) {
   const { metaFields } = settings;
-  return !!metaFields && !!metaFields.userValue.length;
+  return !!metaFields && !_.isEqual(WAZUH_METAFIELDS_SETTINGS, metaFields && metaFields.userValue);
 }
 
-async function updateMetaFieldsSetting(isModified: boolean) {
-  return (
-    !isModified &&
-    (await GenericRequest.request('POST', '/api/kibana/settings', { changes: { metaFields: [] } }))
+async function updateMetaFieldsSetting(isModified:boolean) {
+  return !isModified && await GenericRequest.request(
+    'POST',
+    '/api/kibana/settings',
+    {"changes":{"metaFields":WAZUH_METAFIELDS_SETTINGS}}
   );
 }
